requirements.txt
google-generativeai>=0.3.0
customtkinter>=5.2.0
Pillow>=10.0.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pynput>=1.7.6
python-dotenv>=1.0.0
pyaudio>=0.2.11
ollama>=0.1.0
